refactor: DispatcherServlet 모듈을 app 에서 mvc 로 이동
